acadlink/backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.6.2
python-dotenv==0.19.2
requests==2.28.1